Crate problem080

Problem 80 (Square root digital expansion)

Solution detail

Good approximations to square roots can be found using the convergents to the continued fraction expansion. In particular, if p / q is a convergent to an irrational number x, then the absolute error in this approximation is at most 1 / q^2.

This means that we just need to find a convergent to √n for which the denominator has half as many digits as the required level of accuracy, and then calculate the first decimal digits of this approximation.

Functions

answer
Solve the problem, returning the answer as a String

first_digits_sqrt_sum
Find the sum of the first digits of the square root of the given number.

main

power_of_ten
Finds 10 to the power of n

solve
Find the sum of the first decimal digits of approximations to square roots of the first integers